requirements.txt
absl-py>=2.0.0
attrdict==2.0.1
einops==0.8.0
matplotlib==3.9.2
numpy==1.26.4
pyyaml==6.0.2
torch==1.13.1+cu116
torchvision==0.14.1+cu116
torchaudio==0.13.1
tqdm==4.66.5
geopandas==1.0.1
gpytorch==1.12